[refactor] consolidate email services and templates; remove unused alumni and contact services
