+++ src/services/email/emailServices.ts
@@ -1,5 +1,9 @@
 import { transporter } from '../../config/email';
+import { renderAlumniVerificationEmail } from './templates/alumniDetailsVerification';
+import { renderContactFormEmail } from './templates/contactUs';
 import { renderVerificationEmail } from './templates/verification';
+
+const adminEmails = process.env.EMAIL_ADMINS?.split(',') || [];
 
 interface SendEmailParams {
     to: string;
@@ -28,5 +32,35 @@ export const sendVerificationEmail = async (email: string, token: string) => {
         to: email,
         subject: 'Email Verification',
         html: renderVerificationEmail(verificationUrl),
+    });
+};
+
+export const sendContactUsEmail = async (
+    email: string,
+    name: string,
+    subject: string,
+    message: string,
+) => {
+    return sendEmail({
+        to: adminEmails.join(','),
+        subject: `Contact Form Submission: ${subject}`,
+        html: renderContactFormEmail(name, email, subject, message),
+    });
+};
+
+export const sendAlumniVerificationEmail = async (
+    userId: string,
+    name: string,
+) => {
+    const verificationUrl = `${process.env.ADMIN_URL}/`;
+    const emailHtml = renderAlumniVerificationEmail(verificationUrl, {
+        name,
+        userId,
+    });
+
+    return sendEmail({
+        to: adminEmails.join(','),
+        subject: `Alumni Verification - ${name}`,
+        html: emailHtml,
     });
 };
